Move the zendesk ticket creation route from admin controller to finder controller
We can now remove the admin controller, as this was the last route to move to the finder controller
